# postman/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Learning Journal API",
    "version": "1.0.0",
    "description": "Learning JournalのAPI仕様書"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "開発環境"
    },
    {
      "url": "https://learning-journal.vercel.app",
      "description": "本番環境"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "Unit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "learningGoal": {
            "type": "string",
            "nullable": true
          },
          "preLearningState": {
            "type": "string",
            "nullable": true
          },
          "reflection": {
            "type": "string",
            "nullable": true
          },
          "nextAction": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["not_started", "in_progress", "completed"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "commentsCount": {
            "type": "integer"
          },
          "likesCount": {
            "type": "integer"
          },
          "isLiked": {
            "type": "boolean"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "image": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/units/{id}": {
      "get": {
        "summary": "ユニットの詳細を取得",
        "description": "指定されたIDのユニットの詳細情報を取得します。",
        "tags": ["ユニット"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ユニットID"
          }
        ],
        "responses": {
          "200": {
            "description": "ユニットの詳細取得に成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Unit"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ユニットが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "ユニットを更新",
        "description": "指定されたIDのユニットを更新します。認証が必要です。",
        "tags": ["ユニット"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ユニットID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "タイトル"
                  },
                  "learningGoal": {
                    "type": "string",
                    "description": "学習目標"
                  },
                  "preLearningState": {
                    "type": "string",
                    "description": "事前の学習状態"
                  },
                  "reflection": {
                    "type": "string",
                    "description": "振り返り"
                  },
                  "nextAction": {
                    "type": "string",
                    "description": "次のアクション"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "開始日"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "終了日"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["not_started", "in_progress", "completed"],
                    "description": "ステータス"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "タグ名の配列"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ユニットの更新に成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Unit"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "権限エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "ユニットが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "ユニットを削除",
        "description": "指定されたIDのユニットを削除します。認証が必要です。",
        "tags": ["ユニット"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ユニットID"
          }
        ],
        "responses": {
          "204": {
            "description": "ユニットの削除に成功"
          },
          "401": {
            "description": "認証エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "権限エラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "ユニットが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "サーバーエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}